Update runbook-dynamicgroup-mfa.ps1
Now the script is removing and adding users to the groups, based on the existing group memberships. GroupID Variables are transfered to Automation-Variables.

+++ runbook-dynamicgroup-mfa.ps1
@@ -1,5 +1,5 @@
 <#
-Dynamic Group Runbook - MFA State
+Azure Runbook - Dynamic Group - MFA State
 
 This script is designed for an Azure Runbook to assign users to two Azure AD groups based on their MFA capability (capable / non-capable).
 Before running the runbook, you need to set up an automation account with a managed identity.
@@ -11,24 +11,24 @@ The managed identity requires the following Graph Permissions:
 - UserAuthenticationMethod.Read.All
 
 The script requires the following modules:
+- Az.Accounts (should be installed by default!)
 - Microsoft.Graph.Authentication
 - Microsoft.Graph.Groups
 - Microsoft.Graph.Identity.SignIns
 - Microsoft.Graph.Users
 
-Important: Define the variables for the two necessary groups at the beginning of the script.
+!!! Important: Define the variables for the two necessary groups in the Automation Variables as "dynamicmfa_groupid_capable" and "dynamicmfa_groupid_noncapable", or hardcode them in this script. !!!
 
-Version: 0.1
+Version: 0.2
 Creator: Dominik Gilgen (https://github.com/M365-Consultant)
 Date of creation: 2023-09-22
 License: CC BY-SA 4.0 (Attribution-ShareAlike 4.0 International)
 #>
 
 
-#Define variables (e.g. 3edd8134-e624-451e-abb2-ec82cae1900e):
-$groupid_capable = 'xxx'
-$groupid_noncapable = 'xxx'
-
+#variables (define them on the Automation Variables):
+$groupid_capable = Get-AutomationVariable -Name 'dynamicmfa_groupid_capable'
+$groupid_noncapable = Get-AutomationVariable -Name 'dynamicmfa_groupid_noncapable'
 
 
 #Connect to Microsoft Graph within Azure Automation
@@ -37,36 +37,24 @@ $token = Get-AzAccessToken -ResourceUrl "https://graph.microsoft.com"
 $securetoken = ConvertTo-SecureString -String $token.Token -AsPlainText -Force
 Connect-MgGraph -AccessToken $securetoken
 
-
-# Remove all members from exisiting Group
-$members_mfatrue = Get-MgGroupMember -GroupId $groupid_capable
-$members_mfafalse = Get-MgGroupMember -GroupId $groupid_noncapable
-
-foreach ($user in $members_mfatrue) {
-    Remove-MgGroupMemberByRef -GroupId $groupid_capable -DirectoryObjectId $user.Id
-    }
-
-foreach ($user in $members_mfafalse) {
-    Remove-MgGroupMemberByRef -GroupId $groupid_noncapable -DirectoryObjectId $user.Id
-    }
-
-
 # Adding users to the matching group
 $users = get-mguser -All
+$members_capable = Get-MgGroupMember -GroupId $groupid_capable -All
+$members_noncapable = Get-MgGroupMember -GroupId $groupid_noncapable -All
+
 
 foreach ($user in $users) {
     $mfa = Get-MgUserAuthenticationMethod -UserId $user.UserPrincipalName -All | where {$_.AdditionalProperties."@odata.type" -ne "#microsoft.graph.passwordAuthenticationMethod"}
     if ($mfa.Count -gt 0) {
-    New-MgGroupMember -GroupId $groupid_capable -DirectoryObjectId $user.Id
+            if ($members_capable.Id -notcontains $user.Id){ New-MgGroupMember -GroupId $groupid_capable -DirectoryObjectId $user.Id ; Write-Output $user.UserPrincipalName "added to MFA capable group. User-ID:" $user.Id  }
+            if ($members_noncapable.Id -contains $user.Id){ Remove-MgGroupMemberByRef -GroupId $groupid_noncapable -DirectoryObjectId $user.Id  ; Write-Output $user.UserPrincipalName "removed from MFA non-capable group. User-ID:" $user.Id }
+    }
+    else{
+            if($members_noncapable.Id -notcontains $user.Id){ New-MgGroupMember -GroupId $groupid_noncapable -DirectoryObjectId $user.Id ; Write-Output $user.UserPrincipalName "added to MFA non-capable group. User-ID:" $user.Id }
+            if($members_capable.Id -contains $user.Id){ Remove-MgGroupMemberByRef -GroupId $groupid_capable -DirectoryObjectId $user.Id ; Write-Output $user.UserPrincipalName "removed from MFA capable group. User-ID:" $user.Id }
     }
 }
 
-foreach ($user in $users) {
-    $mfa = Get-MgUserAuthenticationMethod -UserId $user.UserPrincipalName -All | where {$_.AdditionalProperties."@odata.type" -ne "#microsoft.graph.passwordAuthenticationMethod"}
-    if ($mfa.Count -eq 0) {
-    New-MgGroupMember -GroupId $groupid_noncapable -DirectoryObjectId $user.Id
-    }
-}
 
 #Disconnect from Microsoft Graph within Azure Automation
 Disconnect-MgGraph
